scripts/requirements.txt
psycopg2-binary>=2.9.0
redis>=5.0.0
nats-py>=2.7.0
bcrypt>=4.0.0
